[Feature] Static pages rendering 🤩 (#274)
