+++ docker/Dockerfile.prod
@@ -16,9 +16,7 @@ RUN yarn install
 
 COPY . .
 
-RUN yarn build-frontend
-
-RUN yarn compile
+RUN yarn build-all
 
 # Stage 2 - make final image
 FROM node:16.14.0-alpine3.15
@@ -32,4 +30,4 @@ COPY --from=build /usr/src/app/public ./public
 
 ENV NODE_ENV=production
 
-CMD ["node", "dist/bin/server.js"]
+CMD ["node", "dist/backend/server.js"]
