+++ docs-config.yaml
@@ -3,7 +3,7 @@ host: "localhost"
 uploads:
   driver: "local"
   local:
-    path: "./public/uploads"
+    path: "./uploads"
   s3:
     bucket: "my-bucket"
     region: "eu-central-1"
