+++ tsconfig.json
@@ -16,7 +16,7 @@
     // "declarationMap": true,                      /* Generates a sourcemap for each corresponding '.d.ts' file. */
     // "sourceMap": true,                           /* Generates corresponding '.map' file. */
     // "outFile": "./",                             /* Concatenate and emit output to single file. */
-    "outDir": "./dist/",                            /* Redirect output structure to the directory. */
+    "outDir": "./dist/backend",                            /* Redirect output structure to the directory. */
     // "rootDir": "./",                             /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */
     // "composite": true,                           /* Enable project compilation */
     // "tsBuildInfoFile": "./",                     /* Specify file to store incremental compilation information */
@@ -74,12 +74,15 @@
     "skipLibCheck": true,                           /* Skip type checking of declaration files. */
     "forceConsistentCasingInFileNames": true        /* Disallow inconsistently-cased references to the same file. */
   },
-  "include": ["src/**/*"],
+  "include": ["src/backend/**/*"],
   "ts-node": {
     /**
      * Tell ts-node CLI to install the --loader automatically, explained below
      * https://typestrong.org/ts-node/docs/imports/
      */
     "esm": true
-  }
+  },
+  "exclude": [
+    "src/test/**/*"
+  ]
 }
